fix(web-app): use month:'long' for Georgian locale dates throughout
month:'short' with ka-GE falls back to English abbreviations (May, Apr).
month:'long' produces correct Georgian month names (მაისი, აპრილი).

Affects: Briefings list, Calendar item dates, Account formatDateTime,
SubscriptionCard expiry display.

+++ web-app/src/components/SubscriptionCard.tsx
@@ -15,7 +15,7 @@ const PaywallModal = lazy(() =>
 
 function formatDate(iso: string): string {
   const d = new Date(iso);
-  return d.toLocaleDateString('ka-GE', { day: '2-digit', month: 'short', year: 'numeric' });
+  return d.toLocaleDateString('ka-GE', { day: '2-digit', month: 'long', year: 'numeric' });
 }
 
 export function SubscriptionCard() {
